requirements.txt
Flask==2.0.3
pyodbc==4.0.32
python-dotenv==0.20.0
flask-cors==3.0.10

